Fixed psr-0 namespaces
