+++ composer.json
@@ -22,5 +22,8 @@
     },
     "autoload": {
         "psr-0": { "Net_DNS2": "" }
+    },
+    "autoload": {
+        "psr-0": { "Tests_Net_DNS2": "tests/" }
     }
 }
